tests: extract test_io.py from test_masters_recipe.py
utils/io.py had no dedicated test file -- its FileHandler, path builders,
and junk-exclusion tests lived in test_masters_recipe.py. Split them into
test_io.py (77 tests), leaving test_masters_recipe.py with only the recipe
integration + error-path tests (9). TestFilenameConsistency (the
kpf_filepath vs generate_standard_filename contract) moves with the io
tests; CLAUDE.md updated to point at its new home. Pure move.
