+++ CLAUDE.md
@@ -274,7 +274,7 @@ Handler/level configuration lives in exactly one place: `kpfpipe.utils.logger.se
 
 Science: `L0 = {obs_id}.fits` (KPF-native `KP.*`); `L1 = kpf_L1_{YYYYMMDD}T{HHmmss}.fits` — note **no EPRV "S"**, because the EPRV standard defines no L1 (its filename regex only accepts `SL2`/`SL3`/`SL4`); `L2/L4 = kpf_SL{N}_…` (EPRV-standard). Masters (WMKO DRP-RUN-05): `{KOAID-of-first-input}_master_{type}_L{N}.fits`.
 
-Two authorities encode this rule and **must agree per level**: `kpf_filepath(obs_id, level, …)` (`utils/io.py`) is the pipeline's path builder (directory + filename, from an obs_id string) and is what recipes use to write; `<model>.generate_standard_filename()` builds only the basename and is the `to_fits(fn=None)` fallback. `kpf_filepath` itself decomposes into two lower-level helpers it composes: `kpf_directory(obs_id, *, level, data_root, kind)` — the single authority for an **output** directory (`kind` ∈ `science`/`masters`/`QLP`; QLP is `{data_root}/QLP/{datecode}/{obs_id}/{level}`) — and `kpf_filename(obs_id, level, *, master)` — the basename; `kpf_filepath = os.path.join(kpf_directory(…), kpf_filename(…))`. **`kpf_filename` is the single source for the naming rule**: the four science models' `generate_standard_filename` all delegate to `kpf_filename(self.obs_id, level)` (so the object- and string-keyed builders can't drift), which is why every model must carry `obs_id` on every construction path (see the `ORIGID` recovery note under *Header standardization*). Like `check_filename_convention`, **every concrete model declares it explicitly** (L0/L1/L2/L4 delegate to `kpf_filename`; `KPFMasterModel` overrides with the KOAID/MASTYPE master name, since masters carry no `obs_id`), and `KPFDataModel`'s abstract version **raises `NotImplementedError`**. `TestFilenameConsistency` (in `tests/regression/test_masters_recipe.py`) enforces that `generate_standard_filename` and `kpf_filepath` never drift. (Filename *validation* — `check_filename_convention` — is separate and still delegates to rvdata's EPRV check for L2/L4.)
+Two authorities encode this rule and **must agree per level**: `kpf_filepath(obs_id, level, …)` (`utils/io.py`) is the pipeline's path builder (directory + filename, from an obs_id string) and is what recipes use to write; `<model>.generate_standard_filename()` builds only the basename and is the `to_fits(fn=None)` fallback. `kpf_filepath` itself decomposes into two lower-level helpers it composes: `kpf_directory(obs_id, *, level, data_root, kind)` — the single authority for an **output** directory (`kind` ∈ `science`/`masters`/`QLP`; QLP is `{data_root}/QLP/{datecode}/{obs_id}/{level}`) — and `kpf_filename(obs_id, level, *, master)` — the basename; `kpf_filepath = os.path.join(kpf_directory(…), kpf_filename(…))`. **`kpf_filename` is the single source for the naming rule**: the four science models' `generate_standard_filename` all delegate to `kpf_filename(self.obs_id, level)` (so the object- and string-keyed builders can't drift), which is why every model must carry `obs_id` on every construction path (see the `ORIGID` recovery note under *Header standardization*). Like `check_filename_convention`, **every concrete model declares it explicitly** (L0/L1/L2/L4 delegate to `kpf_filename`; `KPFMasterModel` overrides with the KOAID/MASTYPE master name, since masters carry no `obs_id`), and `KPFDataModel`'s abstract version **raises `NotImplementedError`**. `TestFilenameConsistency` (in `tests/regression/test_io.py`) enforces that `generate_standard_filename` and `kpf_filepath` never drift. (Filename *validation* — `check_filename_convention` — is separate and still delegates to rvdata's EPRV check for L2/L4.)
 
 ### Masters Pipeline
